+++ src/main/kotlin/com/petqua/application/auth/AuthDtos.kt
@@ -1,0 +1,6 @@
+package com.petqua.application.auth
+
+data class AuthResponse(
+    val accessToken: String,
+    val refreshToken: String,
+)
