--- src/main/kotlin/com/petqua/domain/auth/OauthTokenInfo.kt
@@ -1,9 +1,0 @@
-package com.petqua.domain.auth
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy
-import com.fasterxml.jackson.databind.annotation.JsonNaming
-
-@JsonNaming(SnakeCaseStrategy::class)
-data class OauthTokenInfo(
-    val accessToken: String
-)
