--- src/main/kotlin/com/petqua/domain/auth/OauthUserInfo.kt
@@ -1,7 +1,0 @@
-package com.petqua.domain.auth
-
-data class OauthUserInfo(
-    val nickname: String,
-    val imageUrl: String,
-    val oauthId: String,
-)
