+++ src/main/kotlin/com/petqua/domain/auth/oauth/OauthClient.kt
@@ -1,4 +1,4 @@
-package com.petqua.domain.auth
+package com.petqua.domain.auth.oauth
 
 import java.net.URI
