+++ src/main/kotlin/com/petqua/domain/auth/oauth/OauthServerType.kt
@@ -1,4 +1,4 @@
-package com.petqua.domain.auth
+package com.petqua.domain.auth.oauth
 
 import com.petqua.common.exception.auth.OauthClientException
 import com.petqua.common.exception.auth.OauthClientExceptionType.UNSUPPORTED_OAUTH_SERVER_TYPE
