--- src/main/kotlin/com/petqua/presentation/OauthResponse.kt
@@ -1,6 +1,0 @@
-package com.petqua.presentation
-
-data class OauthResponse(
-    val accessToken: String,
-    val refreshToken: String,
-)
