+++ src/main/kotlin/com/petqua/presentation/auth/LoginArgumentResolver.kt
@@ -1,13 +1,14 @@
-package com.petqua.domain.auth
+package com.petqua.presentation.auth
 
 import com.petqua.common.exception.auth.AuthException
-import com.petqua.common.exception.auth.AuthExceptionType.INVALID_REFRESH_TOKEN
-import com.petqua.common.exception.auth.AuthExceptionType.INVALID_REQUEST
+import com.petqua.common.exception.auth.AuthExceptionType
+import com.petqua.domain.auth.Accessor
+import com.petqua.domain.auth.Auth
 import com.petqua.domain.auth.token.AuthTokenProvider
 import com.petqua.domain.auth.token.RefreshTokenRepository
 import jakarta.servlet.http.HttpServletRequest
 import org.springframework.core.MethodParameter
-import org.springframework.http.HttpHeaders.AUTHORIZATION
+import org.springframework.http.HttpHeaders
 import org.springframework.stereotype.Component
 import org.springframework.web.bind.support.WebDataBinderFactory
 import org.springframework.web.context.request.NativeWebRequest
@@ -24,6 +25,7 @@ class LoginArgumentResolver(
 
     override fun supportsParameter(parameter: MethodParameter): Boolean {
         return parameter.hasParameterAnnotation(Auth::class.java)
+                && parameter.getParameterType() == Accessor::class.java;
     }
 
     override fun resolveArgument(
@@ -33,19 +35,19 @@ class LoginArgumentResolver(
         binderFactory: WebDataBinderFactory?
     ): Accessor {
         val request = webRequest.getNativeRequest(HttpServletRequest::class.java)
-            ?: throw AuthException(INVALID_REQUEST)
+            ?: throw AuthException(AuthExceptionType.INVALID_REQUEST)
         val refreshToken = request.cookies?.find {it.name == REFRESH_TOKEN_COOKIE}?.value
-        val accessToken = webRequest.getHeader(AUTHORIZATION) as String
+        val accessToken = webRequest.getHeader(HttpHeaders.AUTHORIZATION) as String
         val accessTokenClaims = authTokenProvider.getAccessTokenClaims(accessToken)
         if (refreshToken == null) {
             return Accessor.from(accessTokenClaims)
         }
 
         val savedRefreshToken = refreshTokenRepository.findByMemberId(accessTokenClaims.memberId)
-            ?: throw AuthException(INVALID_REFRESH_TOKEN)
+            ?: throw AuthException(AuthExceptionType.INVALID_REFRESH_TOKEN)
         if (savedRefreshToken.token == refreshToken) {
             return Accessor.from(accessTokenClaims)
         }
-        throw AuthException(INVALID_REFRESH_TOKEN)
+        throw AuthException(AuthExceptionType.INVALID_REFRESH_TOKEN)
     }
 }
