+++ src/main/kotlin/com/petqua/presentation/auth/OauthServerTypeConverter.kt
@@ -1,6 +1,6 @@
-package com.petqua.presentation
+package com.petqua.presentation.auth
 
-import com.petqua.domain.auth.OauthServerType
+import com.petqua.domain.auth.oauth.OauthServerType
 import org.springframework.core.convert.converter.Converter
 
 class OauthServerTypeConverter : Converter<String, OauthServerType> {
