+++ src/test/kotlin/com/petqua/test/config/OauthTestConfig.kt
@@ -1,6 +1,6 @@
 package com.petqua.test.config
 
-import com.petqua.domain.auth.kakao.KakaoOauthApiClient
+import com.petqua.domain.auth.oauth.kakao.KakaoOauthApiClient
 import com.petqua.test.FakeKakaoOauthApiClient
 import org.springframework.boot.test.context.TestConfiguration
 import org.springframework.context.annotation.Bean
